requirements.txt
fastapi
uvicorn
pydantic
slowapi
pyyaml
